Edit readme.md file.
Does not alter any core functionality of the repo. Changes one line in the The Code Structure section.

Before:
In the repo you will find many examples of code which will have the same functionality, however the methods of instantiating Classes and creating new Objects will be different in each piece of code.

After:
In the repo you will find many examples of code which will have the same functionality. However, the methods of instantiating Classes and creating new Objects will be different in each piece of code. 

(I tried taking a screenshot but it wasn't working correctly.)